refactor: Split format package into format/spdx, format/cyclonedx, and format/types
refs: #5
